ClinRAG/requirements.txt
llama-index-core
llama-index-readers-file
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss
faiss-cpu
fastapi
uvicorn
streamlit
pydantic
llama-cpp-python
ragas
